pyproject.toml
[project]
name = "grocery-tracker"
version = "0.1.0"
description = "Grocery receipt analysis using OCR"
requires-python = ">=3.9"
dependencies = [
    "Flask>=2.3.0",
    "Flask-SQLAlchemy>=3.0.0",
    "Flask-Login>=0.6.0",
    "Flask-WTF>=1.1.0",
    "Werkzeug==2.3.7",
    "WTForms>=3.0.0",
    "email-validator>=2.0.0",
    "SQLAlchemy>=2.0.0",
    "easyocr>=1.7.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "scikit-image>=0.21.0",
    "deskew>=0.10.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "altair>=5.0.0",
    "unidecode>=1.3.0",
    "pytz>=2023.3",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "google-genai>=1.47.0",
    "pydantic>=2.12.5",
]

[project.scripts]
grocery-tracker = "g_tracker.wsgi:app"

[tool.uv]
dev-dependencies = []
